% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_inspire.R
\name{add_inspire}
\alias{add_inspire}
\title{add_inspire}
\usage{
add_inspire(
  pts,
  sEPSG,
  iCellSize,
  names_centro = c("x_centroide", "y_centroide")
)
}
\arguments{
\item{pts}{(\strong{df object}) : table of centroids.}

\item{sEPSG}{(\strong{integer or character}) : epsg code.}

\item{iCellSize}{(\strong{integer}) : cells size(s) in meters. Can be a vector for irregular grids}

\item{names_centro}{(\strong{character vector}) : vector of names for longitude/latitude variables}
}
\value{
(\strong{df}) pts table with additional \emph{idInspire} variable
}
\description{
Creates Inpire names for a grid defined with :
\itemize{
\item centroids coordinates
\item squares size(s)
\item Projection system
}

(Ajoute les noms des identifiants en norme Inspire des carreaux définis à partir de :
\itemize{
\item des coordonnées de leurs centroides
\item la taille de leur côté en mètres
\item un système de projection)
}
}
\examples{
pts <- data.frame(x_centroide = c(100, 100, 300, 300, 500), y_centroide = c(100, 300, 100, 300, 100))
add_inspire(pts, sEPSG = 2154, iCellSize = 200)
}
