% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{constituerMatriceEffectifs}
\alias{constituerMatriceEffectifs}
\title{constituerMatriceEffectifs}
\usage{
constituerMatriceEffectifs(vLigneObservation, vColonneObservation)
}
\arguments{
\item{vLigneObservation}{:
\itemize{
\item A \code{vector} containing the line number of each observation
\item (Un \code{vector} contenant le numéro de ligne de chaque observation.)
}}

\item{vColonneObservation}{:
\itemize{
\item A \code{vector} containing the column number of each observation
\item (Un \code{vector} contenant le numéro de colonne de chaque observation.)
}}
}
\value{
Returns a \code{matrix} with the number of observations for each cell.

(Retourne une \code{matrix} avec le nombre d'observations pour chaque cellule.)
}
\description{
Function constituting a membership matrix (Fonction constituant une matrice des effectifs)
}
\author{
\itemize{
\item Psar Analyse Urbaine Insee
\item Arlindo Dos Santos
\item Francois Semecurbe
}
}
