\name{dfPrix_SP95_2016}
\alias{dfPrix_SP95_2016}
\docType{package}
\title{
Unleaded 95 price in France in 2016  (Prix du sans plomb 95 en france en 2016) 
}

\description{
Unleaded 95 price in France in 2016. source : https://www.prix-carburants.gouv.fr/rubrique/opendata/

(Donnees carroyees a 200 metres. source : https://www.prix-carburants.gouv.fr/rubrique/opendata/)
}

\value{
  \item{x}{longitude}
  \item{y}{latitude}
  \item{SP95}{Unleaded price - pris du sans plomb}
 }
 
\examples{
\dontrun{
data(dfPrix_SP95_2016)
}
}
