% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfRestaurantParis.R
\docType{data}
\name{dfRestaurantParis}
\alias{dfRestaurantParis}
\title{Parisian restaurants}
\format{
\subsection{\code{dfRestaurantParis}}{

A data frame with 13,823 rows and 17 columns:
\describe{
\item{x}{longitude}
\item{y}{latitude}
\item{french}{French restaurant}
\item{asian}{asian restaurant}
\item{caribbean}{caribbean restaurant}
\item{maghreb}{maghreb restaurant}
\item{african}{african restaurant}
\item{european}{european restaurant}
\item{south_america}{south_american restaurant}
\item{indian}{indian restaurant}
\item{world}{world food restaurant}
\item{cafeteria}{cafeteria restaurant}
\item{tea_house}{tea_house}
\item{brasserie}{brasserie restaurant}
\item{fastfood}{fastfood restaurant}
\item{bar}{bar}
\item{restaurant}{restaurant}
}
}
}
\source{
source : https://opendata.paris.fr
}
\usage{
dfRestaurantParis
}
\description{
Parisian restaurants (Restaurants parisiens)
}
\keyword{datasets}
