% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfToGrid.R
\name{dfToGrid}
\alias{dfToGrid}
\title{Depreciated function of btb_ptsToGrid}
\usage{
dfToGrid(df, sEPSG, iCellSize = NULL)
}
\arguments{
\item{df}{centroid points}

\item{sEPSG}{epsg code}

\item{iCellSize}{cell size}
}
\value{
a grid
}
\description{
Function dfToGrid has been replaced by btb_ptsToGrid
}
