\name{pixel_france}
\alias{pixel_france}
\docType{package}
\title{
Whole France grid with 1km square tiles (Grille France entière avec des carreaux d'1km de côté) 
}

\description{
Whole France grid with 1km square tiles. Lambert 93 projection 

(Grille France entière avec des carreaux d'1km de côté. Projection Lambert 93)
}

\value{
  \item{x}{longitude}
  \item{y}{latitude}
 }

 
\examples{
\dontrun{
data(pixel_france)
}
}
