% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{constituerGrappes}
\alias{constituerGrappes}
\title{constituerGrappes}
\usage{
constituerGrappes(iNbObsMin, mEffectifs)
}
\arguments{
\item{iNbObsMin}{\itemize{
\item An \code{integer} representing the minimum number of elements in each cluster
\itemize{
\item Un \code{data.frame} représentant le nombre minimum d'éléments dans chaque grappe.)
}
}}

\item{mEffectifs}{\itemize{
\item staffing matrix for each cell (\code{matrix}).
\itemize{
\item matrice des effectifs pour chaque case (\code{matrix}). .)
}
}}
}
\value{
\itemize{
\item Returns a \code{matrix} with the cluster number for each cell .
\item Retourne une \code{matrix} avec le numéro de grappe pour chaque cellule.
}
}
\description{
Quadtree variant function.

(Variante de la fonction Quadtree.)
}
\author{
\itemize{
\item Psar Analyse Urbaine Insee
\item Arlindo Dos Santos
\item Francois Semecurbe
}
}
