% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfPrix_SP95_2016.R
\docType{data}
\name{dfPrix_SP95_2016}
\alias{dfPrix_SP95_2016}
\title{Unleaded 95 price in France in 2016  (Prix du sans plomb 95 en france en 2016)}
\format{
\subsection{\code{dfPrix_SP95_2016}}{

A data frame with 5,573 rows and 3 columns:
\describe{
\item{x}{longitude}
\item{y}{latitude}
\item{SP95}{Unleaded price - prix du sans plomb}
}
}
}
\source{
https://www.prix-carburants.gouv.fr/rubrique/opendata/
}
\usage{
dfPrix_SP95_2016
}
\description{
Unleaded 95 price in France in 2016 (prix du sans-plomb 95 en France en 2016)
}
\keyword{datasets}
