\name{dfRestaurantParis}
\alias{dfRestaurantParis}
\docType{package}
\title{
Parisian restaurants (Restaurants parisiens) 
}

\description{
Parisian restaurants data frame. source : https://opendata.paris.fr

(data frame contenant les restaurants parisiens. source : https://opendata.paris.fr)
}

\value{
  \item{x}{longitude}
  \item{y}{latitude}
  \item{french}{French restaurant}
  \item{asian}{asian restaurant}
 }

\examples{
data(dfRestaurantParis)
}
