% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernelSmoothing.R
\name{kernelSmoothing}
\alias{kernelSmoothing}
\title{kernelSmoothing}
\usage{
kernelSmoothing(...)
}
\arguments{
\item{...}{whatever}
}
\value{
a warning message
}
\description{
Function kernelSmoothing has been replaced by \code{btb_smooth}
}
