% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernelSmoothing.R
\name{kernelSmoothing}
\alias{kernelSmoothing}
\title{Depreciated function of btb_smooth}
\usage{
kernelSmoothing(
  dfObservations,
  sEPSG,
  iCellSize,
  iBandwidth,
  vQuantiles = NULL,
  dfCentroids = NULL,
  fUpdateProgress = NULL,
  iNeighbor = NULL,
  iNbObsMin = 250
)
}
\arguments{
\item{dfObservations}{points}

\item{sEPSG}{epsg code}

\item{iCellSize}{cell size}

\item{iBandwidth}{smoothing bandwidth}

\item{vQuantiles}{quantiles}

\item{dfCentroids}{user grid}

\item{fUpdateProgress}{message parameter}

\item{iNeighbor}{number of neighbors pixels}

\item{iNbObsMin}{technical parameter}
}
\value{
a warning message
}
\description{
Function kernelSmoothing has been replaced by \code{btb_smooth}
}
