% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixel_france.R
\docType{data}
\name{pixel_france}
\alias{pixel_france}
\title{Unleaded 95 price in France in 2016  (Prix du sans plomb 95 en france en 2016)}
\format{
\subsection{\code{pixel_france}}{

A data frame with 54,8495 rows and 2 columns:
\describe{
\item{x}{longitude of pixel's centroids}
\item{y}{latitude of pixel's centroids}
}
}
}
\source{
source : https://opendata.paris.fr
}
\usage{
pixel_france
}
\description{
Whole France grid with 1km square tiles. Lambert 93 projection.

(Grille France entière avec des carreaux d'1km de côté. Projection Lambert 93)
}
\keyword{datasets}
