\name{pixel_france}
\alias{pixel_france}
\docType{package}
\title{
Whole France grid with 1km square tiles (Grille France entiere avec des carreaux d'1km de cote)) 
}

\description{
Whole France grid with 1km square tiles. Lambert 93 projection 

(Grille France entière avec des carreaux d'1km de côté. Projection Lambert 93)
}

\value{
  \item{x}{longitude}
  \item{y}{latitude}
 }

 
\examples{
data(pixel_france_1km_2154)
}
