% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reunion.R
\docType{data}
\name{reunion}
\alias{reunion}
\title{Households of Reunion}
\format{
\subsection{\code{reunion}}{

A data frame with 14,076 rows and 4 columns:
\describe{
\item{x}{longitude}
\item{y}{latitude}
\item{houhold}{number of households - nombre de ménages}
\item{phouhold}{number of poor households - nombre de ménages pauvres}
}
}
}
\source{
Source : Insee, Localized data : Fiscal Revenue - 31/12/2010 and Housing Tax - 01/01/2011

Source : Insee, Revenus Fiscaux Localisés (RFL) au 31 decembre 2010 et Taxe d'habitation (TH) au 1er janvier 2011)
}
\usage{
reunion
}
\description{
Households of Reunion (Menages de La Reunion)
Gridded database with a grid cell resolution of 200 meters. (Données carroyées à 200 mètres).
}
\keyword{datasets}
